Behat API Extension - Initial Commit
This is a extension for Behat to attend a variety of scenarios and
standardize for tests and behaviour validation duraing QA.

There are some open api extension for behat in github, buy anyone
was what we are expecting for our QA validation process.

Basic context and steps for api requests and response validation
are implemented in this extension.

It uses guzzle as client to execute requests.